Converted routing utils to typescript, refactored a bit
